[gwanho] Update Server
이전에 작업하던 내용에서 GameState에 있던 내용들을 TaskManager를 따로 만들어서 관리하도록 유도함
추가적으로 Tick을 구현하기 위해서 준비작업함

+++ GameServer/GameDatabase.cpp
@@ -2,7 +2,7 @@
 #include "GameDatabase.h"
 #include <ConnectionInfo.h>
 
-GameDatabase::GameDatabase(const size_t inThreadPoolSize, const size_t poolSize) : DatabaseManager(inThreadPoolSize, poolSize)
+GameDatabase::GameDatabase(const uint32 inThreadPoolSize, const uint32 poolSize) : DatabaseManager(inThreadPoolSize, poolSize)
 {
 }
